+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## dev
 
+#### What's Improved
+- Tags formatted as `artistssort` or `albumartistssort` are now recognized by Auxio
+
 #### What's Fixed
 - Disc number is no longer mis-aligned when no subtitle is present